+++ packages/workbox-build/package.json
@@ -21,7 +21,6 @@
   "homepage": "https://github.com/GoogleChrome/workbox",
   "dependencies": {
     "babel-runtime": "^6.26.0",
-    "chalk": "^2.1.0",
     "common-tags": "^1.4.0",
     "fs-extra": "^4.0.2",
     "glob": "^7.1.2",
