--- packages/workbox-build/src/lib/constants.js
@@ -1,11 +1,0 @@
-// TODO: Remove me.
-module.exports = {
-  // Max File Size: 2MB
-  maximumFileSize: 2 * 1024 * 1024,
-  defaultGlobPatterns: [
-    '**/*.{js,css,html}',
-  ],
-  defaultGlobIgnores: [
-    'node_modules/**/*',
-  ],
-};
