+++ packages/workbox-build/src/lib/get-composite-details.js
