+++ packages/workbox-build/src/lib/get-string-details.js
