+++ packages/workbox-build/src/lib/get-string-hash.js
