+++ packages/workbox-build/src/lib/no-revision-for-urls-matching-transform.js
@@ -1,4 +1,4 @@
-const errors = require('../errors');
+const errors = require('./errors');
 
 module.exports = (regexp) => {
   if (!(regexp instanceof RegExp)) {
